+++ src/main/java/com/example/entity/ShukakuBoxSum.java
@@ -24,7 +24,7 @@ public class ShukakuBoxSum {
 	@DateTimeFormat(pattern = "yyyy/MM/dd HH:mm:ss")
 	private LocalDateTime registDatetime;
 	private String biko;
-	private int boxSum;  // 収穫ケース数合計
+	private double boxSum;  // 収穫ケース数合計
 	private boolean delflg;
 	
 	/*
