+++ src/main/java/com/example/form/FormKanriMainteHouseDetail.java
@@ -20,18 +20,18 @@ public class FormKanriMainteHouseDetail {
 	private String houseName;
 	
 	//予定収穫ケース数(01月～12月)
-	private int boxSumYTI01 = 0;
-	private int boxSumYTI02 = 0;
-	private int boxSumYTI03 = 0;
-	private int boxSumYTI04 = 0;
-	private int boxSumYTI05 = 0;
-	private int boxSumYTI06 = 0;
-	private int boxSumYTI07 = 0;
-	private int boxSumYTI08 = 0;
-	private int boxSumYTI09 = 0;
-	private int boxSumYTI10 = 0;
-	private int boxSumYTI11 = 0;
-	private int boxSumYTI12 = 0;
+	private double boxSumYTI01 = 0;
+	private double boxSumYTI02 = 0;
+	private double boxSumYTI03 = 0;
+	private double boxSumYTI04 = 0;
+	private double boxSumYTI05 = 0;
+	private double boxSumYTI06 = 0;
+	private double boxSumYTI07 = 0;
+	private double boxSumYTI08 = 0;
+	private double boxSumYTI09 = 0;
+	private double boxSumYTI10 = 0;
+	private double boxSumYTI11 = 0;
+	private double boxSumYTI12 = 0;
 	
 	private String biko;
 	private boolean deleteFlg;
